+++ src/__tests__/stationView.spec.js
@@ -6,6 +6,19 @@ import StationView from '../components/stationView.js'
 
 describe('<StationView>', () => {
   it('should render without crashing', () => {
-    shallow(<StationView />)
+    shallow(<StationView activities={[{ uuid: 'uuid' }]} />)
+  })
+
+  it('should render Activities', () => {
+    const activities = [
+      {
+        uuid: 'uuid1'
+      },
+      {
+        uuid: 'uuid2'
+      }
+    ]
+    const wrapper = shallow(<StationView name='name' description='' activities={activities} />)
+    expect(wrapper.find('.station__activities').children()).toHaveLength(activities.length)
   })
 })