+++ src/utils/api.js
@@ -4,6 +4,10 @@
  * @return {Object} - JS Object from JSON files
  */
 export async function fetchJSON (url, params = {}) {
+  /** MOCK LOAD ACTIVITES */
+  if (url === process.env.PUBLIC_URL + '/json/activities.json') {
+    return mockFetchActivities(params)
+  }
   const request = {
     headers: {
       Accept: 'application/json'
@@ -17,4 +21,24 @@ export async function fetchJSON (url, params = {}) {
     return data.filter(s => params.uuids.includes(s.uuid))
   }
   return data
+}
+
+async function mockFetchActivities ({ uuids }) {
+  const request = {
+    headers: {
+      Accept: 'application/json'
+    },
+    method: 'GET'
+  }
+  const activities = [
+    process.env.PUBLIC_URL + '/json/textActivity.json'
+  ]
+  const resposnes = await Promise.all(
+    activities.map(activityUrl => window.fetch(activityUrl, request))
+  )
+  const data = await Promise.all(
+    resposnes.map(r => r.json())
+  )
+
+  return data.flat().filter(s => uuids.includes(s.uuid))
 }